Section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -44,12 +45,12 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
-   
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto) {
        var auction = _mapper.Map<Auction>(auctionDto);
-       //Todo: add current user as seller 
-       auction.Seller = "test";
+       
+       auction.Seller = User.Identity.Name;
        _context.Auctions.Add(auction);
        var newAuction = _mapper.Map<AuctionDto>(auction);
        await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
@@ -58,13 +59,16 @@ public class AuctionsController : ControllerBase
        return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
     }
 
-
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(UpdateAuctionDto auctionUpdate, Guid id){
         var auction = await _context.Auctions.Include(x => x.Item)
                 .FirstOrDefaultAsync(x => x.Id == id);
+
         if(auction is null) return NotFound();
-        //Todo: check seller == username
+        
+        if(auction.Seller != User.Identity.Name) return Forbid();
+
         auction.Item.Make = auctionUpdate.Make ?? auction.Item.Make;
         auction.Item.Model = auctionUpdate.Model ?? auction.Item.Model;
         auction.Item.Color = auctionUpdate.Color ?? auction.Item.Color;
@@ -77,12 +81,15 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
 
-
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id){
         var auction = await _context.Auctions.FindAsync(id);
+
         if(auction is null) return NotFound();
-        //Todo: check seller == username
+         
+        if(auction.Seller != User.Identity.Name) return Forbid();
+
         _context.Auctions.Remove(auction);
 
         await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
